src/api/requirements.txt
fastapi
uvicorn
sse-starlette
llama_cpp_python 
